Fix bug preventing optimization from firing (#65573)
Summary:
Pull Request resolved: https://github.com/pytorch/pytorch/pull/65573

When we remove mutation on
```
x = [0, 1, 3, 4]
x[-2] = 4
```
we have a safety check that the new index will be in bounds of the old index. in practice, this should always be the case otherwise you would have a runtime error. Within that check (not within the actual adjustment) we were using the wrong length of inputs preventing the optimization from firing.

Test Plan: Imported from OSS

Reviewed By: navahgar

Differential Revision: D31732417

Pulled By: eellison

fbshipit-source-id: dd734254c0212ca459c1c135da262974de5299be