Correct `householder_product` docs. (#68335)
Summary:
Pull Request resolved: https://github.com/pytorch/pytorch/pull/68335

When discussing https://github.com/pytorch/pytorch/pull/63880, we
realised that the docs of `householder_product` were not correct. This
PR fixes this.

The new docs are slightly more difficult, but hopefully correct. Note
that this is a LAPACK function in disguise, so it is expected the
specification to be more difficult than normal.

cc brianjo mruberry jianyuh nikitaved pearu walterddr IvanYashchuk xwang233 Lezcano

Test Plan: Imported from OSS

Reviewed By: mrshenli

Differential Revision: D32429755

Pulled By: mruberry

fbshipit-source-id: 3ac866d30984adcd9f3b83d7fa9ae7b7ae5d4b53